Added support for custom documentation build scripts.
If a module's `package.json` file defines a `build-shared-docs`
script, it will be used to build the documentation for the module.
(See the additions to `README.md` for more details.)

Also made the build process better at ignoring non-module files in the
`modules` and `docs` directories and prettied up its trace output a bit.